allow calculation of actor-specific maximum milk limits
The original formula to calculate the maximum milk limit for all actors
was '(Level+2)*2'. This patch extracts the values and assigns a meaning.
It also introduces a scaling factor for the level based adjustment.

The new formula is:

  (MilkBaseValue + Level*MilkScaleFactor)*BreastCount

Possible future usage:
 - support more than (or less than) 2 breasts
   (may obsolete 'MME.MilkMaid.BreastRows')
 - support manual customization for each actor through debug menu
 - support race-specific milk limits

Currently the values for MilkBaseValue, MilkScaleFactor and BreastCount
are still hardcoded. This will be addressed in later patches.

Something similar can be implemented for other (currently) fixed values to
enhance the roleplaying aspect. (Orcs are more resistant to pain, Khajiit
generate more milk, Elves have lower storage capacity)