extract MME.MilkMaid.PainCount to storage module
The function 'setPainCurrent()' now automatically verifies the provided
value to make sure it is not set to an invalid value. The allowed range
is '0 <= x <= PainMax'. For normal operation this is expected to have
no influence but as a side effect it is now no longer possible to set
a PainCount > PainMax (e.g. up to 100) via the MCM debug menu.
The MCM slider max value has been adjusted to reflect this restriction.

Tracing output is provided if such a condition occurs. Additionally
these functions will return 'true' in case the provided value is
accepted without change or 'false' if the value was outside the
allowed range. (It is safe to ignore these return values.)